+++ test/gen_test.dart
@@ -3,17 +3,28 @@ import 'package:test/test.dart';
 
 import 'test_utils.dart';
 
-const argString = 'pub run build_runner build --delete-conflicting-outputs';
+const argString = 'run build_runner build --delete-conflicting-outputs';
 
 void main() {
-  test('[engine] gen', () async {
-    final result = await testCommand(['gen']);
-    final stdout = result.stdout;
+  test(
+    '[engine] gen',
+    () async {
+      final result = await testCommand(
+        ['gen'],
+        projects: defaultProjects(devDependencies: {'build_runner: any'}),
+        debug: true,
+      );
+      final stdout = result.stdout;
 
-    expect(result.exitCode, isNot(ExitCode.success.code));
+      expect(result.exitCode, ExitCode.success.code);
 
-    expectLine(stdout, ['dart_puby_test', 'dart $argString']);
-    expectLine(stdout, ['flutter_puby_test', 'flutter $argString']);
-    expectLine(stdout, ['fvm_puby_test', 'fvm flutter $argString']);
-  });
+      expectLine(stdout, ['dart_puby_test', 'dart $argString']);
+      expectLine(stdout, ['dart_puby_test/example', 'Skip']);
+      expectLine(stdout, ['flutter_puby_test', 'dart $argString']);
+      expectLine(stdout, ['flutter_puby_test/example', 'Skip']);
+      expectLine(stdout, ['fvm_puby_test', 'fvm dart $argString']);
+      expectLine(stdout, ['fvm_puby_test/example', 'Skip']);
+    },
+    timeout: Timeout.factor(1.5),
+  );
 }